Add tests
Expand the test suite to cover most of the application. Because the
API is pretty exposed at this point, we're not going to test things
like `shakeTimers` fully, as they really should be extracted to a
separate module (or even package? or upstreamed?) and tested
in isolation.

Also return `log` & `unwatch` to `watchWithTime` while maintaining
backwards-compatibility with the normal `log` usage. The API
overall will need to be cleaned up in a future version but this works
well enough.

Fixes #1 & #4.