Add two simple tools for marble diagrams
The first, `parseDiagram`, converts a marble diagram & an object
of events into an array of frames, which is just an array of events
that can be provided to `send`. The second, `sendFrames`, takes
those frames & an `advance` function and iterates over those
frames, sending them down the Observable, and advancing to the
next frame.

These two tools can be combined to integrate marble diagrams &
frames into testing tools.

Fix #2.